Bug 1304972 - Get rid of django-crontab in favor of rq jobs
I replaced the django-crontab jobs with django-admin commands and
wrapped those commands in very thin functions.
These functions are registered in the rq-backed schedule when the app
is loading.
The actual schedule is then stored in a redis sorted set; the
singularity of the jobs in the schedule is guaranteed by setting the job
id to the function name.