Add Dockerfile.dev file
This does everything the main Dockerfile does except COPY’ing
the source code to /app. Instead we’re using a volume to mount it
under /app. We’re also creating /opt/coverage for the coverage
data created when running the tests.

This simplifies the docker-compose set up somewhat since we don’t
require seperation between development and testing environment
anymore that much. The only difference now is that the testing has
DEBUG set to False to more closely match the prod environment.

This also extended the run script to run the given command in case
it didn’t match any of the shortcut names such as “web”, “worker” or
“scheduler” so calls to docker-compose run still work as expected.

The run script will now also wait for the database and redis services
to start before continuing. That is to work around a missing feature
in docker-compose in which startup order isn’t defined.